Move mergeFragmentWithEntity to roosterjs
