+++ packages/roosterjs-content-model/lib/index.ts
@@ -1,9 +1,5 @@
 export { default as domToContentModel } from './publicApi/domToContentModel';
 export { default as contentModelToDom } from './publicApi/contentModelToDom';
-export {
-    default as mergeFragmentWithEntity,
-    preprocessEntitiesFromContentModel,
-} from './publicApi/mergeFragmentWithEntity';
 export { default as insertTable } from './publicApi/table/insertTable';
 export { default as formatTable } from './publicApi/table/formatTable';
 export { default as setTableCellShade } from './publicApi/table/setTableCellShade';
@@ -134,10 +130,7 @@ export {
     ContentModelHandlerTypeMap,
     DefaultImplicitSegmentFormatMap,
 } from './publicTypes/context/ModelToDomSettings';
-export {
-    ModelToDomEntityContext,
-    EntityPlaceholderPair,
-} from './publicTypes/context/ModelToDomEntityContext';
+export { ModelToDomEntityContext } from './publicTypes/context/ModelToDomEntityContext';
 export { ElementProcessor } from './publicTypes/context/ElementProcessor';
 export { ContentModelHandler } from './publicTypes/context/ContentModelHandler';
