+++ packages/roosterjs-content-model/lib/publicTypes/context/ModelToDomEntityContext.ts
@@ -1,17 +1,4 @@
-/**
- * Represent an object pair of Entity DOM node and placeholder comment node
- */
-export interface EntityPlaceholderPair {
-    /**
-     * Wrapper element of element
-     */
-    entityWrapper: HTMLElement;
-
-    /**
-     * Placeholder comment node
-     */
-    placeholder: Comment;
-}
+import { EntityPlaceholderPair } from 'roosterjs-editor-types';
 
 /**
  * Represents context for entity
