+++ packages/roosterjs-editor-types/lib/interface/EntityPlaceholderPair.ts
@@ -1,0 +1,14 @@
+/**
+ * Represent an object pair of Entity DOM node and placeholder comment node
+ */
+export interface EntityPlaceholderPair {
+    /**
+     * Wrapper element of element
+     */
+    entityWrapper: HTMLElement;
+
+    /**
+     * Placeholder comment node
+     */
+    placeholder: Comment;
+}
