+++ packages/roosterjs-editor-types/lib/interface/index.ts
@@ -119,3 +119,4 @@ export {
     ImageSelectionRange,
     SelectionRangeEx,
 } from './SelectionRangeEx';
+export { EntityPlaceholderPair } from './EntityPlaceholderPair';
